Move data extracts to new folder structure
